+++ changelog.md
@@ -17,13 +17,20 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
-  * [Releases v1.0.1](#releases-v101)
+  * [Releases v1.1.0](#Releases-v110)
+  * [Releases v1.0.1](#Releases-v101)
   * [Releases v1.0.0](#releases-v100)
 
 ---
 ---
 
 ## Changelog
+
+### Releases v1.1.0
+
+1. Add example [PWM_manual](https://github.com/khoih-prog/ESP32_FastPWM/tree/main/examples/PWM_manual) to demo how to correctly use PWM to generate waveform. Check [About DCValue in setPWM_manual #2](https://github.com/khoih-prog/AVR_PWM/discussions/2)
+2. Add function `setPWM_DCPercentage_manual()` to facilitate the setting PWM DC manually by using `DCPercentage`, instead of `absolute DCValue` depending on varying resolution
+3. Fix glitch when changing PWM frequency. Check [Frequency change on cycle end. #1](https://github.com/khoih-prog/ESP32_FastPWM/discussions/1)
 
 ### Releases v1.0.1